Prevent everything from being deleted if updateIds is empty array

+++ server/thanks/collection.ts
@@ -95,7 +95,9 @@ class ThanksCollection {
    * @param {Types.ObjectId[] | string[]} updateIds - The id of the updates
    */
    static async deleteManybyUpdateIds(updateIds: Types.ObjectId[] | string[]): Promise<void> {
-    await ThanksModel.deleteMany({ updateId: { $in: updateIds } });
+    if (updateIds.length) {
+      await ThanksModel.deleteMany({ updateId: { $in: updateIds } });
+    }
   }
 
   /**
